LF-15757 timings: revert STAR 5543305 workaround
STAR 5543305 workaround was implemented in PHYINIT 2024.09 in order
to fix a potential PPT2 failure if dfi_lp_data_req is enabled,
STAR fix leading to an average 30% power increase on DDR rail.
DDRC however has no PPT2 support on iMX95, 943 and 952, so it's
safe revert the changes introduced by STAR 5543305 workaround
in order to restore the previous power levels.

Signed-off-by: Florin Leotescu <florin.leotescu@nxp.com>
Signed-off-by: Adrian Stoica <adrian.stoica@nxp.com>
Signed-off-by: Michael Kjar <michael.kjar@@nxp.com>
Signed-off-by: Viorel Suman <viorel.suman@nxp.com>